Split multiple envs tests
