+++ OmniGibson/tests/utils.py
@@ -248,3 +248,102 @@ def remove_all_systems(scene):
     for system in scene.active_systems.values():
         system.remove_all_particles()
     og.sim.step()
+
+
+# ---------------------------------------------------------------------------
+# Shared helpers for test_multiple_envs_*.py
+# ---------------------------------------------------------------------------
+
+MULTI_ENV_ROBOTS = ["fetch", "tiago", "r1", "r1pro"]
+
+MULTI_ENV_TASK_TYPES = ["DummyTask", "PointNavigationTask", "PointReachingTask", "GraspTask"]
+
+GRASP_OBJECTS_CFG = [
+    {
+        "type": "PrimitiveObject",
+        "name": "grasp_obj",
+        "primitive_type": "Cube",
+        "rgba": [1.0, 0, 0, 1.0],
+        "scale": [0.04, 0.04, 0.04],
+        "mass": 0.1,
+        "position": [0.5, 0.0, 0.55],
+    }
+]
+
+ROBOT_JOINT_COUNTS = {"fetch": 8, "tiago": 8, "r1": 10, "r1pro": 11}
+
+
+def _grasp_reset_pose_path(robot):
+    """Create a temporary precached reset pose file sized for the given robot."""
+    import json
+    import tempfile
+
+    n_joints = ROBOT_JOINT_COUNTS[robot]
+    reset_poses = [{"joint_pos": [0.0] * n_joints, "base_pos": [0.0, 0.0, 0.0], "base_ori": [0.0, 0.0, 0.0, 1.0]}]
+    f = tempfile.NamedTemporaryFile(mode="w", suffix=".json", delete=False)
+    json.dump(reset_poses, f)
+    f.close()
+    return f.name
+
+
+def multi_env_task_cfg(task_type, robot="fetch"):
+    """Return the task config dict for a given task type string."""
+    if task_type == "GraspTask":
+        return {
+            "type": "GraspTask",
+            "obj_name": "grasp_obj",
+            "objects_config": GRASP_OBJECTS_CFG,
+            "termination_config": {"max_steps": 10},
+            "precached_reset_pose_path": _grasp_reset_pose_path(robot),
+        }
+    return {"type": task_type}
+
+
+def multi_env_progress(test_name, counter):
+    """Print progress for the current test."""
+    counter["current"] += 1
+    n = counter["current"]
+    total = counter["total"]
+    print(f"\n{'='*60}")
+    print(f"[{n}/{total}] RUNNING: {test_name}")
+    print(f"{'='*60}")
+
+
+def multi_env_passed(test_name):
+    """Print pass confirmation."""
+    print(f"[PASSED] {test_name}")
+
+
+def _init_multi_env_macros():
+    """Set simulator macros (only once, before first Environment is created)."""
+    if og.sim is None:
+        gm.RENDER_VIEWER_CAMERA = False
+        gm.ENABLE_OBJECT_STATES = True
+        gm.USE_GPU_DYNAMICS = True
+        gm.ENABLE_FLATCACHE = False
+        gm.ENABLE_TRANSITION_RULES = False
+    else:
+        og.sim.stop()
+
+
+def setup_multi_environment(num_of_envs, robot="fetch", task_type="DummyTask", additional_objects_cfg=None):
+    """Create an Environment with *num_of_envs* cloned scenes."""
+    _init_multi_env_macros()
+
+    cfg = {
+        "env": {"num_envs": num_of_envs},
+        "scene": {
+            "type": "InteractiveTraversableScene",
+            "scene_model": "Rs_int",
+            "load_object_categories": ["floors", "walls"],
+        },
+        "robots": [{"model": robot, "obs_modalities": []}],
+        "task": multi_env_task_cfg(task_type, robot=robot),
+    }
+    if additional_objects_cfg:
+        cfg["objects"] = additional_objects_cfg
+
+    print(f"  Setting up env: num_envs={num_of_envs}, robot={robot}, task={task_type}")
+    env = og.Environment(configs=cfg)
+    print(f"  Environment created successfully")
+    return env
